Another attempt at fixing CaseOfTheMissingTweet
Twitter is really having issues with database consistency
when replying quickly with tweets. This time, they are returning
the same 385 error when trying to post a tweet. However, unlike before
this time the tweet actually is getting created, even though twitter
returns an error.

The attempted solution to this is to crawl through our own timeline
looking for the tweet we posted, and then continuing on.

Added log statements to verify this works or doesn't next time twitter
decides to behave this way